requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
ConfigArgParse==1.1
Flask==1.1.1
gevent==1.5a3
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
locustio==0.14.5
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.0
python-engineio==3.12.1
python-socketio==4.5.1
pyzmq==19.0.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0